# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: CrowdTune API
  version: 0.0.0
  description: Phase 8-9 skeleton. /healthz public, /api/* requires Neon Auth bearer JWT.
  license:
    name: Sustainable Use License v1.0
    url: ../../LICENSE.md
servers:
  - url: http://localhost:8080
    description: Local dev
tags:
  - name: meta
    description: Server meta and health
  - name: identity
    description: Authenticated user info
  - name: parties
    description: Party creation and metadata
  - name: spotify
    description: Spotify OAuth token management
paths:
  /healthz:
    get:
      operationId: getHealth
      tags: [meta]
      summary: Liveness and DB connectivity check
      security: []
      responses:
        '200':
          description: Server is alive and dependencies are reachable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '503':
          description: Server is up but a critical dependency is down
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/parties:
    post:
      operationId: createParty
      tags: [parties]
      summary: Create a new party
      description: >
        Creates a new party with a server-generated 6-character alphanumeric short code.
        The authenticated user becomes the party host.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePartyRequest'
      responses:
        '201':
          description: Party created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Party'
        '400':
          description: Request body is missing or has invalid fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Bearer token missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/parties/{partyId}:
    get:
      operationId: getParty
      tags: [parties]
      summary: Get party metadata by short code
      security:
        - bearerAuth: []
      parameters:
        - name: partyId
          in: path
          required: true
          schema:
            type: string
          description: The 6-character party short code
      responses:
        '200':
          description: Party metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Party'
        '404':
          description: Party not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/me:
    get:
      operationId: getMe
      tags: [identity]
      summary: Return the authenticated user's identity claims
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User identity extracted from the bearer JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: Bearer token missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/spotify/token:
    post:
      operationId: postSpotifyToken
      tags: [spotify]
      summary: Exchange a Spotify authorization code for an access token
      description: >
        Accepts the short-lived Spotify OAuth authorization code and redirect URI,
        exchanges them with Spotify's token endpoint, encrypts the refresh token,
        and persists credentials server-side. Returns the access token and metadata.
        The refresh token is never returned to the client.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpotifyTokenExchangeRequest'
      responses:
        '200':
          description: Spotify access token exchanged successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpotifyTokenExchangeResponse'
        '400':
          description: Request body is missing or has invalid fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Bearer token missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Spotify token endpoint returned an error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error (encryption or DB failure)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Neon Auth (Better-Auth) JWT. Validated via JWKS at NEON_AUTH_JWKS_URL.
  schemas:
    Party:
      type: object
      required: [id, shortId, name, hostUserId, settings, isActive, createdAt, updatedAt]
      properties:
        id:
          type: string
          description: 6-character alphanumeric party short code
        shortId:
          type: string
          description: Alias for id; redundant field for frontend compatibility
        name:
          type: string
          description: Display name of the party
        hostUserId:
          type: string
          description: JWT sub claim of the party host
        settings:
          type: object
          description: Opaque party settings blob
          additionalProperties: true
        isActive:
          type: boolean
          description: Whether the party is currently active
        createdAt:
          type: string
          format: date-time
          description: ISO 8601 creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: ISO 8601 last-updated timestamp
    HealthResponse:
      type: object
      required: [status, db]
      properties:
        status:
          type: string
          enum: [ok]
        db:
          type: string
          enum: [ok, down]
        version:
          type: string
          description: API version (semver)
    MeResponse:
      type: object
      required: [user_id]
      properties:
        user_id:
          type: string
          description: Subject claim from the JWT (Neon Auth user ID)
        email:
          type: string
          format: email
        role:
          type: string
    CreatePartyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          description: Display name for the party (non-empty, whitespace-trimmed)
        settings:
          type: object
          description: Opaque party settings blob stored verbatim as JSONB
          additionalProperties: true
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error message
    SpotifyTokenExchangeRequest:
      type: object
      required: [code, redirectUri]
      properties:
        code:
          type: string
          description: The one-time Spotify OAuth authorization code from the redirect URL
        redirectUri:
          type: string
          description: The redirect URI that was sent to Spotify during the authorization request
    SpotifyTokenExchangeResponse:
      type: object
      required: [accessToken, expiresIn, scopes]
      properties:
        accessToken:
          type: string
          description: Spotify OAuth access token (valid approximately one hour)
        expiresIn:
          type: integer
          description: Seconds until the access token expires
        scopes:
          type: string
          description: Space-delimited Spotify OAuth scopes granted